Refactor SSE test
